backend/requirements.txt
# -----------------------------------------------------------------
# FastAPI 및 서버 관련 패키지
# -----------------------------------------------------------------
fastapi==0.116.1
uvicorn[standard]==0.35.0
gunicorn==23.0.0

# -----------------------------------------------------------------
# 데이터베이스 및 ORM 관련 패키지
# -----------------------------------------------------------------
SQLAlchemy==2.0.42
psycopg2-binary==2.9.10
alembic==1.16.4

# -----------------------------------------------------------------
# 데이터 유효성 검사 및 설정 관리
# -----------------------------------------------------------------
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1

# -----------------------------------------------------------------
# 인증 및 보안 관련 패키지
# -----------------------------------------------------------------
passlib
bcrypt<4.0.0
python-jose[cryptography]==3.5.0
tzdata==2025.2